requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
blinker==1.9.0
brotli==1.2.0
certifi==2026.2.25
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
ffmpy==1.0.0
filelock==3.25.2
Flask==3.1.3
fsspec==2026.2.0
gradio==6.10.0
gradio_client==2.4.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
text-unidecode==1.3
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
Werkzeug==3.1.6
